+++ src/app/app-routing.module.ts
@@ -11,8 +11,11 @@ const routes: Routes = [
     path: 'home',
     loadChildren: () => import('./modules/+home/home.module').then( m => m.HomeModule)
   },
+  {
+    path: 'characters',
+    loadChildren: () => import('./modules/+characters/characters.module').then( m => m.CharactersModule)
+  },
 ];
-
 
 
 @NgModule({