+++ src/app/modules/+characters/characters-main/characters-main.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-characters-main',
+  templateUrl: './characters-main.component.html',
+})
+export class CharactersMainComponent {
+
+}