+++ src/app/modules/+characters/characters-routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { CharactersMainComponent } from './characters-main/characters-main.component';
+
+
+const routes: Routes = [
+  {
+    path: '',
+    component: CharactersMainComponent
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class CharactersRoutinModule {}