+++ src/app/modules/+characters/characters.module.ts
@@ -1,0 +1,26 @@
+import { HttpClient } from "@angular/common/http";
+import { NgModule } from "@angular/core";
+import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { CharactersMainComponent } from "./characters-main/characters-main.component";
+import { CharactersRoutinModule } from "./characters-routing.module";
+
+export const createTranslateLoader = (http: HttpClient) =>
+  new TranslateHttpLoader(http, './assets/i18n/', '.json');
+
+@NgModule({
+  declarations: [
+    CharactersMainComponent
+  ],
+  imports: [
+    CharactersRoutinModule,
+    TranslateModule.forChild({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: createTranslateLoader,
+        deps: [HttpClient]
+      }
+    })
+  ],
+})
+export class CharactersModule { }