+++ src/app/modules/+home/home/home.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
+import { DB_TABLES } from 'src/app/shared/constants/db.constants';
+import { Character } from 'src/app/shared/models/character.model';
 import { DatabaseService } from 'src/app/shared/services/db.service';
 
 @Component({
@@ -10,16 +12,11 @@ export class HomeComponent implements OnInit {
 
   constructor(
     private translate: TranslateService,
-    private dbService: DatabaseService,
+    private db: DatabaseService
     ) { }
 
-  async ngOnInit() {
-    this.setTranslationsLanguage('es');
-    this.dbService.createDatabase();
+  ngOnInit(): void {
+
   }
 
-  private setTranslationsLanguage(language: string): void {
-    this.translate.setDefaultLang(language);
-    this.translate.use(language);
-  }
 }