+++ src/app/shared/constants/db.constants.ts
@@ -1,5 +1,10 @@
 export const DB_NAME = 'RQDB';
 
 export const DB_DEFINITION = {
-  PJS: '++id, name, age',
+  PJS: '++id, name, race, age, culture, profesion, bornIn, country, religions',
 };
+
+export enum DB_TABLES {
+  PJS = 'PJS',
+}
+