+++ src/app/shared/controllers/character-creation.controller.ts
@@ -1,0 +1,3 @@
+export const getUniqueID = (name: String) => {
+  return name + '_' + new Date().getTime();
+};